YOLOv8
The occupancy detecting program is created using the SOTA YOLOv8 model. The model was trained on 1,300 images with their respective sizes and annotations in 100 epochs. This program is trained to work on any parking lot. However, the user will have to obtain the coordinates of the slots manually, using the 'draw.py' program. In this case, the "slots.json" file consists of the coordinates of the parking lot slots in "carPark.mp4".